Extend Deduping to better handle previous transactions.
When a transaction has been reconciled and ingested successfully via the
ynab.py, it's ynab-id is attached to its beanfile entry. However, any transfer-
or subtransaction-ID's are not.

When subsequent calls to ynab.py are made, that ynab-id is added to the set of
seen_transactions but the transfer- and subtransaction-ids are missing.

This raises an issue where an account reconcilled 2 accounts, reconcilled at
different times, may not be linked and hence double-entered into the beanfile.

This commit duplicates the deduping code used to capture transfer- and
subtransaction-ids into the seen_transactions set when the stream of YNAB
transactions is downloaded and processed.

This is fragile as it will only properly dedup transactions which are included
in the YNAB transaction fetch process. If the --since flag is used, beware!